Refactor: kis_test -> data
