+++ auto_stock/data/admin.py
