+++ auto_stock/data/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class KisTestConfig(AppConfig):
+class DataConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'kis_test'
+    name = 'data'
