+++ auto_stock/data/models.py
