+++ auto_stock/data/tests.py
