+++ auto_stock/data/urls.py
@@ -1,10 +1,10 @@
 from django.urls import path
 
-from kis_test import views
+from data import views
 
 
 urlpatterns = [
-    # this urls are included under the prefix: /api/kis-test/
+    # this urls are included under the prefix: /api/data/
     path("price/", views.RealtimeQuoteView.as_view(), name="kis-realtime-quotes"), # websocket (OK)
     path("daily/", views.DailyPriceView.as_view(), name="kis-daily-price"),        # restapi   (OK)
     path("token/", views.TokenStatusView.as_view(), name="kis-auth-token-status"), # restapi   (OK)
