+++ auto_stock/data/views.py
